# apps/docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spillway API",
    "version": "1.0.0",
    "description": "The control plane for AI spend. One OpenAI-compatible gateway (`/v1/*`) plus a governance and administration API (`/api/*`). Request bodies are generated from the zod schemas in `@spillway/shared` — the single source of truth."
  },
  "servers": [
    {
      "url": "https://api.spillway.dev",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "virtualKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Virtual key `mk-live-…` on the data plane. `x-api-key: mk-live-…` is also accepted."
      },
      "sessionToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "WorkOS session access token (human dashboard sessions) on the control plane."
      },
      "adminKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin API key `mka-…` for machine control-plane access. Requires `X-Spillway-Org`."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "budget_exceeded",
                  "rate_limited",
                  "model_not_allowed",
                  "key_not_found",
                  "key_revoked",
                  "key_paused",
                  "rule_deny",
                  "all_providers_failed",
                  "upstream_error",
                  "invalid_request",
                  "internal_error",
                  "unauthenticated",
                  "forbidden",
                  "not_found",
                  "conflict",
                  "validation_error",
                  "org_required",
                  "last_owner",
                  "key_expired",
                  "service_unavailable",
                  "no_candidates",
                  "provider_key_decrypt_failed",
                  "provider_unavailable",
                  "upstream_parse_error",
                  "request_too_large",
                  "model_not_found",
                  "no_route_available",
                  "approval_required",
                  "tier_required",
                  "plan_upgrade_required",
                  "cel_parse_error",
                  "cel_type_error",
                  "cel_cost_exceeded",
                  "cel_ast_too_large",
                  "cel_banned_macro",
                  "cel_regex_too_long",
                  "approval_chain_unsatisfiable",
                  "not_pending",
                  "self_approval_not_allowed",
                  "not_an_approver",
                  "unknown_effect",
                  "invalid_action_token",
                  "threshold_condition_not_isolated"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ],
        "description": "Control-plane error envelope (§1.4)."
      },
...
      "ApprovalPolicyList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": {}
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "has_more": {
                "type": "boolean"
              },
              "next_cursor": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "has_more",
              "next_cursor"
            ]
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "description": "Cursor-paginated list (§1.3)."
      },
...
    },
...
  },
  "paths": {
...
    "/api/approval-policies": {
      "get": {
        "tags": [
          "Approvals"
        ],
        "summary": "List approval policies.",
        "security": [
          {
            "sessionToken": []
          },
          {
            "adminKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "601c5d19-0b88-42f7-acda-d1d31abd100d"
            },
            "required": false,
            "name": "X-Spillway-Org",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalPolicyList"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Approvals"
        ],
        "summary": "Create an approval policy.",
        "security": [
          {
            "sessionToken": []
          },
          {
            "adminKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "601c5d19-0b88-42f7-acda-d1d31abd100d"
            },
            "required": false,
            "name": "X-Spillway-Org",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "kind": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 60
                  },
                  "scopeType": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 30
                  },
                  "scopeId": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  },
                  "definition": {
                    "type": "object",
                    "properties": {
                      "tiers": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "min_amount_usd": {
                              "type": "string",
                              "pattern": "^\\d+(\\.\\d{1,6})?$"
                            },
                            "auto_approve": {
                              "type": "boolean"
                            },
                            "steps": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "approvers": {
                                    "type": "object",
                                    "properties": {
                                      "roles": {
                                        "type": "array",
                                        "items": {
                                          "type": "string"
                                        }
                                      },
                                      "user_ids": {
                                        "type": "array",
                                        "items": {
                                          "type": "string"
                                        }
                                      }
                                    }
                                  },
                                  "quorum": {
                                    "anyOf": [
                                      {
                                        "type": "string",
                                        "enum": [
                                          "all"
                                        ]
                                      },
                                      {
                                        "type": "string",
                                        "enum": [
                                          "any"
                                        ]
                                      },
                                      {
                                        "type": "integer",
                                        "minimum": 1
                                      }
                                    ]
                                  },
                                  "notify_only": {
                                    "type": "boolean"
                                  }
                                },
                                "required": [
                                  "approvers",
                                  "quorum"
                                ]
                              }
                            }
                          },
                          "required": [
                            "min_amount_usd"
                          ]
                        },
                        "minItems": 1
                      },
                      "expiry_hours": {
                        "type": "integer",
                        "exclusiveMinimum": 0
                      }
                    },
                    "required": [
                      "tiers"
                    ]
                  },
                  "enabled": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "name",
                  "kind",
                  "definition"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {},
                  "description": "Resource object."
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/approval-policies/{id}": {
      "patch": {
        "tags": [
          "Approvals"
        ],
        "summary": "Update an approval policy.",
        "security": [
          {
            "sessionToken": []
          },
          {
            "adminKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "601c5d19-0b88-42f7-acda-d1d31abd100d"
            },
            "required": false,
            "name": "X-Spillway-Org",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "definition": {
                    "type": "object",
                    "properties": {
                      "tiers": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "min_amount_usd": {
                              "type": "string",
                              "pattern": "^\\d+(\\.\\d{1,6})?$"
                            },
                            "auto_approve": {
                              "type": "boolean"
                            },
                            "steps": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "approvers": {
                                    "type": "object",
                                    "properties": {
                                      "roles": {
                                        "type": "array",
                                        "items": {
                                          "type": "string"
                                        }
                                      },
                                      "user_ids": {
                                        "type": "array",
                                        "items": {
                                          "type": "string"
                                        }
                                      }
                                    }
                                  },
                                  "quorum": {
                                    "anyOf": [
                                      {
                                        "type": "string",
                                        "enum": [
                                          "all"
                                        ]
                                      },
                                      {
                                        "type": "string",
                                        "enum": [
                                          "any"
                                        ]
                                      },
                                      {
                                        "type": "integer",
                                        "minimum": 1
                                      }
                                    ]
                                  },
                                  "notify_only": {
                                    "type": "boolean"
                                  }
                                },
                                "required": [
                                  "approvers",
                                  "quorum"
                                ]
                              }
                            }
                          },
                          "required": [
                            "min_amount_usd"
                          ]
                        },
                        "minItems": 1
                      },
                      "expiry_hours": {
                        "type": "integer",
                        "exclusiveMinimum": 0
                      }
                    },
                    "required": [
                      "tiers"
                    ]
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {},
                  "description": "Resource object."
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Approvals"
        ],
        "summary": "Delete an approval policy.",
        "security": [
          {
            "sessionToken": []
          },
          {
            "adminKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "601c5d19-0b88-42f7-acda-d1d31abd100d"
            },
            "required": false,
            "name": "X-Spillway-Org",
            "in": "header"
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}